Added support for conf.json options passing, with entry jd3t, and the option sourceBaseDir.
If sourcebaseDir is set, relative paths for source files, will be
searched for in this directory. The current behaviour is to use
env.dirname, which fails if the sources is not actually placed inside
the jsdoc directory (because jsdoc appears to set env.dirname to this
directory, at least in my setup).

If sourceBaseDir is not set, env.dirname is used as default, which
matches current behaviour.

Example of use: With the following conf.json:
{
  // ...
  "templates": {
    "jd3t" : { "sourceBaseDir" : "/home/jenkins/src" }
  }
}
relative sources passed to jsdoc will be looked for in
/home/jenkins/src when the template runs.

Because you also need to pass the sources relative to jsdoc, you will
most often change to sourceBaseDir right before running jsdoc. Then
everything should work.