Changes to output of deprecated tag.
This commit modifies the output of the deprecated tag to use same css
styles as current JSDoc3 default template, and, when applicable,
change a true bool into "Yes". Also added a space in the output
between the word deprecated and any optional text. This conforms to
the description of the deprecated tag at
http://usejsdoc.org/tags-deprecated.html

Changed default value to be bold and red (as current JSDoc default
template), by using the same style class "important".

Sorry about the tabs and spaces mixup in this and previous commits -
will be setting up the editor to use tabs from this point forward...
:-(